+++ src/parser/inline/autolink/anchor.test.ts
@@ -10,12 +10,13 @@ describe('Unit: parser/inline/autolink/anchor', () => {
       assert.deepStrictEqual(inspect(parser('')), undefined);
       assert.deepStrictEqual(inspect(parser('>')), undefined);
       assert.deepStrictEqual(inspect(parser('>>')), undefined);
-      assert.deepStrictEqual(inspect(parser(' >0')), undefined);
+      assert.deepStrictEqual(inspect(parser('>>>')), undefined);
+      assert.deepStrictEqual(inspect(parser(' >>0')), undefined);
     });
 
     it('valid', () => {
-      assert.deepStrictEqual(inspect(parser('>0')), [['<a class="anchor" href="?res=0">&gt;0</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('>0aA')), [['<a class="anchor" href="?res=0aA">&gt;0aA</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('>>0')), [['<a class="anchor" href="?res=0">&gt;&gt;0</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('>>0aA')), [['<a class="anchor" href="?res=0aA">&gt;&gt;0aA</a>'], '']);
     });
 
   });